requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
nltk>=3.6.0
yfinance>=0.1.70
beautifulsoup4>=4.9.0
requests>=2.25.0
python-dateutil>=2.8.0
